Put contract back at workspace root again
