--- integration-tests/public/runtime-call-contract/flipper/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "flipper"
-version = "5.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-publish = false
-
-[dependencies]
-ink = { path = "../../../../crates/ink", default-features = false }
-flipper-traits = { path = "../traits", default-features = false }
-
-[lib]
-path = "lib.rs"
-
-[features]
-default = ["std"]
-std = [
-    "ink/std",
-]
-ink-as-dependency = []
